# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecuritySuite (secapi) — APIs propias",
    "version": "0.1.0",
    "description": "Catálogo generado por `pnpm docs:api` a partir de `src/app/api/**/route.ts`.\n\nLo generado describe **lo que el código hace hoy**, incluida la autenticación real de\ncada handler. Las descripciones a mano viven en `docs/api/anotaciones.yaml` y ganan\nsobre lo inferido; `GET /api/docs/spec` devuelve el documento ya mergeado.\n\nNo editar este archivo a mano: se regenera."
  },
  "servers": [
    {
      "url": "http://localhost:4005",
      "description": "desarrollo local (pnpm dev)"
    },
    {
      "url": "https://secapi.riogas.com.uy",
      "description": "secapi (PM2, puerto 3001)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por `POST /api/db/login`. Payload: `{ iss, username, userId, sistema }` — no lleva el flag root."
      },
      "cookieToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "El mismo JWT, tomado de la cookie `token` cuando no hay header."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
    "/api/db/accesos": {
      "get": {
        "tags": [
          "accesos"
        ],
        "summary": "GET /api/db/accesos?usuarioId=&aplicacionId=&funcionalidadId=",
        "description": "GET /api/db/accesos?usuarioId=&aplicacionId=&funcionalidadId=",
        "operationId": "get_api_db_accesos",
        "parameters": [
          {
            "name": "aplicacionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "funcionalidadId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "usuarioId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/accesos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "accesos"
        ],
        "summary": "POST /api/db/accesos — upsert de un acceso individual",
        "description": "POST /api/db/accesos — upsert de un acceso individual\nBody: { usuarioId, funcionalidadId, efecto?, creadoEn?, fechaDesde?, fechaHasta? }",
        "operationId": "post_api_db_accesos",
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/accesos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "accesos"
        ],
        "summary": "DELETE /api/db/accesos?usuarioId=&funcionalidadId=",
        "description": "DELETE /api/db/accesos?usuarioId=&funcionalidadId=",
        "operationId": "delete_api_db_accesos",
        "parameters": [
          {
            "name": "funcionalidadId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "usuarioId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/accesos/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
    "/api/db/acciones": {
      "get": {
        "tags": [
          "acciones"
        ],
        "summary": "GET /api/db/acciones?funcionalidadId=&estado=&search=",
        "description": "GET /api/db/acciones?funcionalidadId=&estado=&search=",
        "operationId": "get_api_db_acciones",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "funcionalidadId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/acciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "acciones"
        ],
        "summary": "Body: { nombre, descripcion?, estado?, funcionalidadIds?: number[] }",
        "description": "POST /api/db/acciones\nBody: { nombre, descripcion?, estado?, funcionalidadIds?: number[] }",
        "operationId": "post_api_db_acciones",
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/acciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
  },
...
}